+++ backend/entities/session.go
