+++ backend/entities/session_test.go
