+++ public/index.html
@@ -6,10 +6,10 @@
 
 	<title>Ochi</title>
 
-	<link rel='stylesheet' href='/global.css'>
-	<link rel='stylesheet' href='/build/bundle.css'>
+	<link rel='stylesheet' href='/src/global.css'>
+	<link rel='stylesheet' href='/src/build/bundle.css'>
 
-	<script defer src='/build/bundle.js'></script>
+	<script defer src='/src/build/bundle.js'></script>
 	<script src="https://accounts.google.com/gsi/client" async defer></script>
 </head>
